requirements.txt
# Core Dependencies
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.30
langchain-google-genai==2.1.10
langchain-postgres==0.0.15
langchain-text-splitters==0.3.9

# Database & Vector Store
psycopg2-binary==2.9.10
pgvector==0.3.6

# PDF Processing
pypdf==6.0.0

# Environment & Config
python-dotenv==1.1.1

# HTTP & API clients
openai==1.102.0
requests==2.32.5

# Data Processing
numpy==1.26.4